add readOnly
